[#33] feat: complete roulette organized component
