+++ src/components/roulette/iconRoulette.tsx
@@ -1,19 +1,11 @@
 'use client';
 
 import Icons from '@/components/common/icons/icons';
-import React from 'react';
+import React, { useMemo } from 'react';
 // eslint-disable-next-line
 import _ from 'lodash';
 
 export default function MenuRoulette({ componentKey }: { componentKey: number }) {
-  const shuffleArray = (array: String[]) => {
-    const newArray = [...array];
-    for (let i = array.length - 1; i > 0; i -= 1) {
-      const j = Math.floor(Math.random() * (i + 1));
-      [newArray[i], newArray[j]] = [newArray[j], newArray[i]];
-    }
-    return newArray;
-  };
   function getBgColorClass(index: number) {
     const colorClasses = {
       0: 'bg-roulette_item_bg1',
@@ -24,18 +16,26 @@ export default function MenuRoulette({ componentKey }: { componentKey: number })
     // @ts-ignore
     return colorClasses[index % 4];
   }
-  const menuItems = shuffleArray([
-    'chicken',
-    'cake',
-    'frenchFries',
-    'hamburger',
-    'hotdog',
-    'pizza',
-    'kebab',
-    'noodle',
-    'seshi',
-  ]);
-  const extendedMenuItems = _.flatMap(_.times(2, () => menuItems));
+  const shuffleArray = useMemo(() => {
+    const array = [
+      'chicken',
+      'cake',
+      'frenchFries',
+      'hamburger',
+      'hotdog',
+      'pizza',
+      'kebab',
+      'noodle',
+      'seshi', // 여기서 'seshi'는 오타일 수 있으니, 의도한 단어로 교체하세요.
+    ];
+    const newArray = [...array];
+    for (let i = array.length - 1; i > 0; i -= 1) {
+      const j = Math.floor(Math.random() * (i + 1));
+      [newArray[i], newArray[j]] = [newArray[j], newArray[i]];
+    }
+    return newArray;
+  }, []);
+  const extendedMenuItems = _.flatMap(_.times(2, () => shuffleArray));
 
   return (
     <div
