+++ src/components/roulette/menuList.ts
@@ -1,5 +1,5 @@
 const menuList = [
-  '세종감자탕',
+  '결과는..?',
   '바다회사랑',
   '봉구비어',
   '피자나라치킨공주',
